clean up example types stories
